users-service/src/main/java/com/example/usersservice/controller/AuthenticationController.java
package com.example.usersservice.controller;

import com.example.usersservice.service.AuthenticationService;
import com.example.usersservice.model.LoginDTO;
import com.example.usersservice.model.RegisterDTO;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/v1/auth")
@CrossOrigin("*")
public class AuthenticationController {
    private final AuthenticationService authenticationService;

    public AuthenticationController(AuthenticationService authenticationService) {
        this.authenticationService = authenticationService;
    }

    @PostMapping("/register")
    public ResponseEntity<?> register(@Valid @RequestBody RegisterDTO body){
        return authenticationService.register(body);
    }
    @PostMapping("/login")
    public ResponseEntity<?> login(@Valid @RequestBody LoginDTO body){
        return authenticationService.login(body);
    }
}
